Remove wildcard from variable sourcing
Resolves the issue where `join_buffer%` matches both `join_buffer_size` and `join_buffer_space_limit`, which in turn results in an error like:

```
./tuning-primer.sh: line 402: export: `2097152': not a valid identifier
```

This was originally necessary because older version of MySQL would use `join_buffer` as the variable instead of `join_buffer_size`.  If you are using a version that old, I apologize for breaking your script, but the good news is, you're probably already broken in dozens of other ways running decade-old code, so I don't feel too bad about piling on.