fix(#8): Add support for MySQL 8 (sort of)
It doesn't take long reading https://stackoverflow.com/questions/3836214
to abandon all hope for humanity.

The relevant line here is
https://github.com/mysql/mysql-server/blob/8e797a5d6eb3a87f16498edcb7261a75897babae/mysys_ssl/my_getopt.cc#L658

The reason why it all goes sideways is because of this wonderful set of
lines:
https://github.com/BMDan/tuning-primer.sh/blob/a998f5f3c4cd62410bf581691b04d3d3310af70b/tuning-primer.sh#L185-L208

The warning emitted by my_getopt.cc causes these to evaluate false.  Of
course, that wouldn't be so bad if I could just suppress the warning.
Lots of ways to do that, but all of them suck at least a little for
debugging down the road, and more importantly, they'd require me to
either globally disable warnings (er... no) or do lots of global
output-munging.  I'm not willing to do either.  So I do the only thing I
can think of that doesn't suffer from those downsides: even if the user
doesn't want to, I store their credentials.

Now, I warn the user about this.  It's behind `mktemp`, which means only
the user, themselves, should be able to see those credentials ever
again.  But consider yourself warned; at the end of the day, depending
on your configuration, you *might* be persisting credentials to disk.

Closes #8.