Add support for mariadb and mariadb-admin (#40)
This adds support for using mariadb and mariadb-admin, if they are available on the system.
It avoids so the deprecation warning of mysqladmin in newer mariadb versions:

```
# mysqladmin ping
mysqladmin: Deprecated program name. It will be removed in a future release, use '/usr/bin/mariadb-admin' instead
mysqld is alive
```

This warning prevents the script to run, since a following test on the output of mysqladmin fails.